Add files via upload
added "used" and "loop" check boxes aswell as a sleep timer to enter commands mode allowing you to have it some what act like key presser and enter a command every X(ms) or uncheck it to have it only enter it once the trigger key is pressed. added the ability to have all 4 key presser keys to different F keys or all the same. remember to activate the trigger keys if you change them or it will stay on the old key